simulation_parameters as optional parameter in add_stage() in mfp5io.py
Made 3 changes to add_stage() function to add simulation parameters to a simulation stage:

1) additional but optional simulation_paramters parameter as a dictionary)
2) Updated docstring
3) logic to add stage_parameters as stage attributes, including None and empty check

Next up is pylmps to be modified to provide a dictionary containing stage parameters to the add_stage() function.

Pros:
- Shouldn´t break code
- Parameters from pylmps.MD_init() should be easy to put into the stages if the code over there is changed.

Cons: How do we have to change code in pylmps and other parts that stuff like external potentials etc. are also put into the stage attributes?